client: modularized Step 1 component for ease of access
